+++ src/testing/gameboard.test.js
@@ -1,6 +1,19 @@
 import Gameboard from '../containers/gameboard';
 
-const gameboard = Gameboard();
+// const gameboard = Gameboard();
+let gameboard;
+
+beforeEach(() => {
+  gameboard = Gameboard();
+  gameboard.placeShip([5, 3], false);
+  gameboard.placeShip([2, 8], true);
+
+  gameboard.receiveAttack([5, 3]);
+  gameboard.receiveAttack([6, 3]);
+
+  gameboard.receiveAttack([2, 8]);
+  gameboard.receiveAttack([2, 7]);
+});
 
 describe(`Tests if gameboard.board is a 10x10 grid`, () => {
   test(`Gameboard has 10 rows`, () => {
@@ -13,42 +26,48 @@ describe(`Tests if gameboard.board is a 10x10 grid`, () => {
 });
 
 describe(`Tests gameboard.placeShip`, () => {
-  gameboard.placeShip([5, 3], false);
-  const shipCoordinates_00 = [
-    gameboard.board[7][4],
-    gameboard.board[7][5],
-    gameboard.board[7][6],
-    gameboard.board[7][7],
-    gameboard.board[7][8],
-  ];
-  test(`A ship of length 5 is placed at [5, 3] horizontally : gameboard.placeShip([5, 3])`, () => {
+  test(`A ship of length 3 is placed at [5, 3] horizontally: gameboard.placeShip([5, 3])`, () => {
+    const shipCoordinates_00 = [
+      gameboard.board[7][4],
+      gameboard.board[7][5],
+      gameboard.board[7][6],
+    ];
     expect(shipCoordinates_00.every((coordinate) => coordinate !== null)).toBeTruthy();
   });
 
-  gameboard.placeShip([2, 8], true);
-  const shipCoordinates_01 = [gameboard.board[2][1], gameboard.board[3][1], gameboard.board[4][1]];
-  test(`A ship of length 3 is placed at [2, 8] horizontally : gameboard.placeShip([2, 8])`, () => {
+  test(`A ship of length 3 is placed at [2, 8] horizontally: gameboard.placeShip([2, 8])`, () => {
+    const shipCoordinates_01 = [
+      gameboard.board[2][1],
+      gameboard.board[3][1],
+      gameboard.board[4][1],
+    ];
     expect(shipCoordinates_01.every((coordinate) => coordinate !== null)).toBeTruthy();
+  });
+
+  test(`There is already a ship at [5, 3] horizontally: gameboard.placeShip([5, 3])`, () => {
+    expect(() => gameboard.placeShip([5, 3], false)).toThrow(Error);
+  });
+
+  test(`There must be at least 1 empty coordinate between ships`, () => {
+    expect(() => gameboard.placeShip([5, 2], false)).toThrow(Error);
+  });
+
+  test(`There must be at least 1 empty coordinate between ships`, () => {
+    expect(() => gameboard.placeShip([4, 4], false)).toThrow(Error);
+  });
+
+  test(`There must be at least 1 empty coordinate between ships`, () => {
+    expect(() => gameboard.placeShip([3, 2], false)).toThrow(Error);
   });
 });
 
 describe(`Tests gameboard.receiveAttack`, () => {
   test(`gameboard.receiveAttack([5, 3]) will hit a ship`, () => {
-    gameboard.receiveAttack([5, 3]);
-    gameboard.receiveAttack([6, 3]);
-    gameboard.receiveAttack([7, 3]);
-    gameboard.receiveAttack([8, 3]);
-
     expect(gameboard.hitShots).toContainEqual([5, 3]);
     expect(gameboard.hitShots).toContainEqual([6, 3]);
-    expect(gameboard.hitShots).toContainEqual([7, 3]);
-    expect(gameboard.hitShots).toContainEqual([8, 3]);
   });
 
   test(`gameboard.receiveAttack([2, 8]) will hit a ship`, () => {
-    gameboard.receiveAttack([2, 8]);
-    gameboard.receiveAttack([2, 7]);
-
     expect(gameboard.hitShots).toContainEqual([2, 8]);
     expect(gameboard.hitShots).toContainEqual([2, 7]);
   });
@@ -65,12 +84,12 @@ describe(`Tests gameboard.receiveAttack`, () => {
 });
 
 describe(`Tests ship.isSunk`, () => {
-  test(`Tests if ship at [5, 3] to [9, 3] is sunk`, () => {
+  test(`Tests if ship at [5, 3] to [8, 3] is sunk`, () => {
     expect(gameboard.board[7][4].isSunk()).toBeFalsy();
   });
 
-  test(`Tests if ship at [5, 3] to [9, 3] is sunk`, () => {
-    gameboard.receiveAttack([9, 3]);
+  test(`Tests if ship at [5, 3] to [8, 3] is sunk`, () => {
+    gameboard.receiveAttack([7, 3]);
     expect(gameboard.board[7][4].isSunk()).toBeTruthy();
   });
 
@@ -80,11 +99,12 @@ describe(`Tests ship.isSunk`, () => {
 });
 
 describe(`Tests gameboard.getStatus`, () => {
-  test(`gameboard.getStatus() should return false`, () => {
+  test(`Not all ships on the board have been sunk, gameboard.getStatus() should return false`, () => {
     expect(gameboard.getStatus()).toBeFalsy();
   });
 
-  test(`gameboard.getStatus() should return true`, () => {
+  test(`All ships on the board have been sunk, gameboard.getStatus() should return true`, () => {
+    gameboard.receiveAttack([7, 3]);
     gameboard.receiveAttack([2, 6]);
     expect(gameboard.getStatus()).toBeTruthy();
   });